Move stage monitoring to new observability module.
