+++ server/server/data_loader/tests/test_arangoload.py
@@ -6,7 +6,7 @@ import pytest
 from common.arangodb import get_db, delete_db
 from common.utils import current_app
 from data_loader import arangoload
-from data_loader.arangoload import StagePrinter
+from data_loader.observability import StagePrinter
 from migrations.runner import run_migrations
 
 
@@ -32,7 +32,7 @@ def test_do_collect_data_stage(data_load_app):
         git_repository = data_load_app.config.get('DATA_REPO')
         arangoload.collect_data(data_dir, git_repository)
 
-@pytest.mark.skip('Long running test.')
+# @pytest.mark.skip('Long running test.')
 def test_do_entire_run(data_load_app):
     with data_load_app.app_context():
         db = get_db()
